requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.2
Pillow==10.4.0
requests==2.32.3